+++ src/base/host/host_matrix_dia.cpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2018-2023 Advanced Micro Devices, Inc. All rights Reserved.
+ * Copyright (C) 2018-2024 Advanced Micro Devices, Inc. All rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -27,6 +27,7 @@
 #include "../../utils/log.hpp"
 #include "../matrix_formats_ind.hpp"
 #include "host_conversion.hpp"
+#include "host_io.hpp"
 #include "host_matrix_csr.hpp"
 #include "host_vector.hpp"
 
@@ -259,6 +260,40 @@ namespace rocalution
         }
 
         return false;
+    }
+
+    template <typename ValueType>
+    bool HostMatrixDIA<ValueType>::ReadFileRSIO(const std::string& filename)
+    {
+        int64_t nrow;
+        int64_t ncol;
+        int64_t nnz;
+        int64_t ndiag;
+
+        int*       offset = NULL;
+        ValueType* val    = NULL;
+
+        if(read_matrix_dia_rocsparseio(nrow, ncol, nnz, ndiag, &offset, &val, filename.c_str())
+           != true)
+        {
+            return false;
+        }
+
+        this->Clear();
+        this->SetDataPtrDIA(&offset, &val, nnz, nrow, ncol, ndiag);
+
+        return true;
+    }
+
+    template <typename ValueType>
+    bool HostMatrixDIA<ValueType>::WriteFileRSIO(const std::string& filename) const
+    {
+        return write_matrix_dia_rocsparseio(this->nrow_,
+                                            this->ncol_,
+                                            this->mat_.num_diag,
+                                            this->mat_.offset,
+                                            this->mat_.val,
+                                            filename.c_str());
     }
 
     template <typename ValueType>